niffler_ui_tests/pyproject.toml
[project]
name = "niffler-e-2-e-tests-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Evgeniy Volochek",email = "volochek93@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "playwright (>=1.54.0,<2.0.0)",
    "faker (>=37.5.3,<38.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
